requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
channels==4.2.2
channels_redis==4.2.1
Django==5.1.7
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-extra-fields==3.7.0
drf-yasg==1.21.10
environs==14.1.1
filetype==1.2.0
frozenlist==1.5.0
idna==3.10
inflection==0.5.1
marshmallow==3.26.1
msgpack==1.1.1
multidict==6.2.0
packaging==24.2
pillow==11.1.0
propcache==0.3.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==6.2.0
sqlparse==0.5.3
uritemplate==4.1.1
yarl==1.18.3
